% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayGenes.R
\docType{data}
\name{c2}
\alias{c2}
\title{C2}
\format{
A data frame with 148489 rows and 2 columns.
}
\source{
c2.cp.v7.5.1.symbols.gmt
}
\usage{
data(c2)
}
\description{
This Gene sets are derived from diverse sources,
such as online pathway databases and biomedical literature,
and are curated by individual domain experts.
Each gene set page specifies the source of the corresponding gene set.
The C2 collection is categorized into two subcollections:
Chemical and Genetic Perturbations (CGP) and Canonical Pathways (CP).
}
