% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpath}
\alias{cpath}
\title{Find a pathway between two matrix items}
\usage{
cpath(x, a, b, l)
}
\arguments{
\item{x}{A numeric matrix.}

\item{a}{Integer. Index of observation.}

\item{b}{Integer. Index of observation.}

\item{l}{Integer. Length of pathway.}
}
\description{
Find a pathway between two matrix items
}
