% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway.R
\name{DEGsData}
\alias{DEGsData}
\title{DEGsData}
\usage{
DEGsData(file, EntrezID = NULL, log2FC = NULL)
}
\arguments{
\item{file}{the file path and name, it should be a csv file.}

\item{EntrezID}{If using Symbol name or Ensembl ID, please convert them first.}

\item{log2FC}{log2 Fold Change}
}
\value{
a gene list ranked by log2FC
}
\description{
Load differentially expressed genes and rank the genes by log2FC.
}
\details{
Different Expression Genes Data Prepare
}
