% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayGenes.R
\docType{data}
\name{go.bp}
\alias{go.bp}
\title{GO.BP}
\format{
A data frame with 631855 rows and 2 columns.
}
\source{
c5.go.bp.v7.5.1.symbols.plusCAF_Matrisome.gmt
}
\usage{
data(c2)
}
\description{
Biological Process from Gene Ontology (GO)
}
