% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayGenes.R
\docType{data}
\name{go.cc}
\alias{go.cc}
\title{GO.CC}
\format{
A data frame with 99098 rows and 2 columns.
}
\source{
c5.go.cc.v7.5.1.symbols.gmt
}
\usage{
data(c2)
}
\description{
Cellular Component (CC) from Gene Ontology (GO)
}
