% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayGenes.R
\docType{data}
\name{go.mf}
\alias{go.mf}
\title{GO.MF}
\format{
A data frame with 108789 rows and 2 columns.
}
\source{
c5.go.mf.v7.5.1.symbols.plusCAF_Matrisome.gmt
}
\usage{
data(c2)
}
\description{
Molecular Function (MF) from Gene Ontology (GO)
}
