% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway.R
\name{GSEA_bubble}
\alias{GSEA_bubble}
\title{GSEA Bubble Plot}
\usage{
GSEA_bubble(data, outpath)
}
\arguments{
\item{data}{A data frame containing the required columns: 'ID',
'enrichmentScore', 'up_down', 'classification', 'GeneRatio', 'count',
and 'pvalue'.}

\item{outpath}{The output path where the bubble plot images will be saved.}
}
\value{
Saves the bubble plot images to the specified output path.
}
\description{
This function creates a GSEA bubble plot for each specified classification
('C2', 'GO.BP', 'GO.CC', 'GO.MF') and saves it to the provided output path.
}
