% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway.R
\name{GSEA_histogram}
\alias{GSEA_histogram}
\title{GSEA Histogram Plotting}
\usage{
GSEA_histogram(data, outpath)
}
\arguments{
\item{data}{A data frame containing the following columns: 'ID',
'enrichmentScore', 'up_down', and 'classification'.}

\item{outpath}{A string representing the output path where
the histogram plot will be saved.}
}
\description{
This function generates a GSEA histogram plot using the input data
and saves it to the specified output path.
}
