% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{navigate}
\alias{navigate}
\alias{navigate_ordered}
\alias{navigate_unique}
\alias{navigate_any}
\title{Limit which points are included in the nearest neighbor search}
\usage{
navigate(.p, ...)

navigate_ordered(x, pi, p1, p2)

navigate_unique(x, pi, p1, p2)

navigate_any(x, pi, p1, p2)
}
\arguments{
\item{.p}{A function that returns a vector of indices}

\item{...}{Additional arguments passed on to \code{.p}}

\item{x}{The original matrix}

\item{pi}{Indices of already-selected-nodes}

\item{p1}{Start and end points used in \link{pathway}}

\item{p2}{Start and end points used in \link{pathway}}
}
\value{
Integer. An integer vector of indices within \code{x} forming the search
space for the next step of the nearest neighbor search.
}
\description{
When seeking the next point in a space, search only among those points that
fulfill a given function based on the previously-selected points.
}
\details{
The supplied functions \link{navigate_ordered}, \link{navigate_unique}, and
\link{navigate_any} can be supplied directly to \link{pathway.}

\link{navigate} is a generator that takes an arbitrary user-supplied predicate
function that will generate a vector of search indices when given \code{x}, \code{pi},
\code{p1}, and \code{p2}. This can be useful for drawing a path that, for example, only
moves forward in time, or which introduces a selection probability based on
attributes of the point supplied in another table.
}
\section{Functions}{
\itemize{
\item \code{navigate_ordered}: Select only points with indices following the ones
already selected.

\item \code{navigate_unique}: Select only points that have not yet been visited.

\item \code{navigate_any}: Select any point, even ones that have already been
visited.
}}

