% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway}
\alias{pathway}
\title{Find a pathway of nearest neighbors between two points in a multidimensional
matrix}
\usage{
pathway(x, p1, p2, n = 4L, navigator = navigate_unique, ...,
  verbose = FALSE)
}
\arguments{
\item{x}{A two-dimensional numeric matrix.}

\item{p1}{Integer. Row index of point at the start of the path.}

\item{p2}{Integer. Row index of point at the end of the path.}

\item{n}{Integer. Number of intervening points to find.}

\item{navigator}{Function for constraining nearest neighbor search based on
previously-selected steps.. See \link{navigate} for more details.}

\item{...}{Additional arguments passed to \link[distances:distances]{distances::distances}.}

\item{verbose}{Display progress messages.}
}
\value{
A list with the following values
\itemize{
\item \code{line} A matrix of \code{n} ideal points along a line between \code{p1} and \code{p2}.
\item \code{i} Row indices of \code{x} indicating matched points.
\item \code{p1}, \code{p2} Inidices of points origially passed in.
}
}
\description{
Given two points in a matrix \code{p1} and \code{p2}, find the points nearest to an
ideal path of \code{n} points between them.
}
\details{
To put restrictuions on the search made, specify a \code{navigator} function.
Several are supplied in this package, such as \link{navigate_unique}, which will
produce a path that does not revisit any points, or \link{navigate_ordered}, that
will on return points from progressively later posiiton int he matrix. It is
also possible to specify a custom function using \link{navigate}.
}
\examples{
set.seed(10)
m <- matrix(runif(1000), nrow = 500, ncol = 2)
m[2,]
m[5,]
pathway(m, 2, 11)

pathway(m, 2, 400, navigate = navigate_ordered)
}
