% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pathway}
\alias{plot_pathway}
\title{Plot calculated pathway over original data}
\usage{
plot_pathway(m, p)
}
\arguments{
\item{m}{Original matrix.}

\item{p}{The results of \link{matrix_pathway}.}
}
\description{
Plot calculated pathway over original data
}
