% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pathway.R
\name{matrix_pathway}
\alias{matrix_pathway}
\title{Find a pathway of nearest neighbors between two points in a multidimensional
matrix}
\usage{
matrix_pathway(x, p1, p2, n = 4L, k = 1L, ...)
}
\arguments{
\item{x}{A two-dimensional numeric matrix.}

\item{p1}{Integer. Row index of point at the start of the path.}

\item{p2}{Integer. Row index of point at the end of the path.}

\item{n}{Integer. Number of intervening points to find.}

\item{k}{Integer. Number of nearest neighbors to locate for each point on the
ideal path, passed to \link[distances:nearest_neighbor_search]{distances::nearest_neighbor_search} If \code{k > 1} then
one real point per \code{n} will be randomly sampled from the \code{k} points
returned by the search.}

\item{...}{Additional arguments passed to \link[distances:distances]{distances::distances}.}
}
\value{
A list with the following values
\itemize{
\item \code{line} \code{n} points along a line between \code{p1} and \code{p2}.
\item \code{i} Row indices of \code{x} indicating matched points.
\item \code{ni} Indices of nearest neighbors in \code{x} found for for each point in \code{line}.
\item \code{p1}, \code{p2} Inidices of points origially passed in.
}
}
\description{
Given two points in a matrix \code{p1} and \code{p2}, find the points nearest to an
ideal path between them.
}
\details{
By default \code{k} is 1, and will return the closest possible points to the ideal
line between \code{p1} and \code{p2}. Setting k to a higher number will mean points are
sampled from a larger range.
}
\examples{
set.seed(10)
m <- matrix(runif(1000), nrow = 500, ncol = 2)
m[2,]
m[5,]
matrix_pathway(m, 2L, 11L, k = 1L)
}
