% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway}
\alias{pathway}
\title{Find a pathway of nearest neighbors between two points in a multidimensional
matrix}
\usage{
pathway(x, p1, p2, n = 4L, navigator = navigate_unique, ...,
  verbose = FALSE)
}
\arguments{
\item{x}{A two-dimensional numeric matrix.}

\item{p1}{Integer. Row index of point at the start of the path.}

\item{p2}{Integer. Row index of point at the end of the path.}

\item{n}{Integer. Number of intervening points to find.}

\item{navigator}{Function for constraining nearest neighbor search based on
previously-selected steps.. See \link{navigate} for more details.}

\item{...}{Additional arguments passed to \link[distances:distances]{distances::distances}.}

\item{verbose}{Display progress messages.}
}
\value{
A list with the following values
\itemize{
\item \code{line} \code{n} points along a line between \code{p1} and \code{p2}.
\item \code{i} Row indices of \code{x} indicating matched points.
\item \code{p1}, \code{p2} Inidices of points origially passed in.
}
}
\description{
Given two points in a matrix \code{p1} and \code{p2}, find the points nearest to an
ideal path between them.
}
\details{
By default \code{k} is 1, and will return the closest possible points to the ideal
line between \code{p1} and \code{p2}. Setting k to a higher number will mean points are
sampled from a larger range.
}
\examples{
set.seed(10)
m <- matrix(runif(1000), nrow = 500, ncol = 2)
m[2,]
m[5,]
pathway(m, 2, 11)
}
